Some clean up in o.a.c.util Use generics, remove unused code, fix various Eclipse warnings.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@695273 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/util/MD5Encoder.java
@@ -58,8 +58,8 @@ public final class MD5Encoder {
         char[] buffer = new char[32];
 
         for (int i=0; i<16; i++) {
-            int low = (int) (binaryData[i] & 0x0f);
-            int high = (int) ((binaryData[i] & 0xf0) >> 4);
+            int low = binaryData[i] & 0x0f;
+            int high = (binaryData[i] & 0xf0) >> 4;
             buffer[i*2] = hexadecimal[high];
             buffer[i*2 + 1] = hexadecimal[low];
         }
